+++ fiboa_cli/datasets/template.py
@@ -31,11 +31,18 @@ ATTRIBUTION = "© 2024 ABC Corp."
 # 1. a SPDX license identifier (including "dl-de/by-2-0" / "dl-de/zero-2-0"), or
 LICENSE = "CC-BY-4.0"
 # 2. a STAC Link Object with relation type "license"
-# LICENSE = {'title': 'CC-BY-4.0', 'href': 'https://creativecommons.org/licenses/by/4.0/', 'type': 'text/html', 'rel': 'license'}
+# LICENSE = {"title": "CC-BY-4.0", "href": "https://creativecommons.org/licenses/by/4.0/", "type": "text/html", "rel": "license"}
 
 # Map original column names to fiboa property names
+# You also need to list any column that you may have added in the MIGRATION function (see below).
 COLUMNS = {
-    'area_m': 'area'
+    "area_m": "area"
+}
+
+# Add columns with constant values.
+# The key is the column name, the value is a constant value that's used for all rows.
+ADD_COLUMNS = {
+    "determination_datetime": "2020-01-01T00:00:00Z"
 }
 
 # A list of implemented extension identifiers
@@ -47,7 +54,7 @@ EXTENSIONS = []
 # Function signature:
 #   func(column: pd.Series) -> pd.Series
 COLUMN_MIGRATIONS = {
-    'area_m': lambda column: column * 0.0001
+    "area_m": lambda column: column * 0.0001
 }
 
 # Filter columns to only include the ones that are relevant for the collection,
@@ -66,10 +73,10 @@ MIGRATION = None
 # Schemas for the fields that are not defined in fiboa
 # Keys must be the values from the COLUMNS dict, not the keys
 MISSING_SCHEMAS = {
-    'required': ['my_id'], # i.e. non-nullable properties
-    'properties': {
-        'my_id': {
-            'type': 'string'
+    "required": ["my_id"], # i.e. non-nullable properties
+    "properties": {
+        "my_id": {
+            "type": "string"
         }
     }
 }
@@ -85,14 +92,15 @@ def convert(output_file, cache_file = None, source_coop_url = None, collection =
     1. Run global migration (if provided through MIGRATION)
     2. Run filters to remove rows that shall not be in the final data
        (if provided through COLUMN_FILTERS)
-    3. Run column migrations (if provided through COLUMN_MIGRATIONS)
-    4. Duplicate columns (if an array is provided as the value in COLUMNS)
-    5. Rename columns (as provided in COLUMNS)
-    6. Remove columns (if column is not present as value in COLUMNS)
-    7. Create the collection
-    8. Change data types of the columns based on the provided schemas
+    3. Add columns with constant values
+    4. Run column migrations (if provided through COLUMN_MIGRATIONS)
+    5. Duplicate columns (if an array is provided as the value in COLUMNS)
+    6. Rename columns (as provided in COLUMNS)
+    7. Remove columns (if column is not present as value in COLUMNS)
+    8. Create the collection
+    9. Change data types of the columns based on the provided schemas
     (fiboa spec, extensions, and MISSING_SCHEMAS)
-    9. Write the data to the Parquet file
+    10. Write the data to the Parquet file
 
     Parameters:
     output_file (str): Path where the Parquet file shall be stored.
@@ -116,6 +124,7 @@ def convert(output_file, cache_file = None, source_coop_url = None, collection =
         source_coop_url=source_coop_url,
         extensions=EXTENSIONS,
         missing_schemas=MISSING_SCHEMAS,
+        column_additions=ADD_COLUMNS,
         column_migrations=COLUMN_MIGRATIONS,
         column_filters=COLUMN_FILTERS,
         migration=MIGRATION,