+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- ...
+- `fiboa convert`: Added step that allows to set constant values (`ADD_COLUMNS`)
 
 ### Changed
 
-- ...
+- `fiboa convert`: `determination_datetime` is not required any longer
 
 ### Deprecated